Implementar ferramenta para migrar dados, em vez de migrar pelo reload
